backend/requirements.txt
asgiref==3.8.1
boto3==1.34.109
botocore==1.34.109
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
django-countries==7.6.1
django-storages==1.14.3
django-templated-mail==1.1.1
django-vote==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.7
jmespath==1.0.1
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
